refactor(dev): remove "Plugin" suffix from shared plugins
