+++ packages/remix-dev/compiler/browserjs/compiler.ts
@@ -11,8 +11,8 @@ import type { CompileOptions } from "../options";
 import { browserRouteModulesPlugin } from "./plugins/routes";
 import { browserRouteModulesPlugin as browserRouteModulesPlugin_v2 } from "./plugins/routes_unstable";
 import { cssFilePlugin } from "../plugins/cssImports";
-import { deprecatedRemixPackagePlugin } from "../plugins/deprecatedRemixPackagePlugin";
-import { emptyModulesPlugin } from "../plugins/emptyModulesPlugin";
+import { deprecatedRemixPackagePlugin } from "../plugins/deprecatedRemixPackage";
+import { emptyModulesPlugin } from "../plugins/emptyModules";
 import { mdxPlugin } from "../plugins/mdx";
 import { urlImportsPlugin } from "../plugins/urlImportsPlugin";
 import { cssBundleUpdatePlugin } from "./plugins/cssBundleUpdate";
