+++ packages/remix-dev/compiler/css/compiler.ts
@@ -13,8 +13,8 @@ import type { CompileOptions } from "../options";
 // import { browserRouteModulesPlugin } from "../plugins/browserRouteModulesPlugin";
 // import { browserRouteModulesPlugin as browserRouteModulesPlugin_v2 } from "../plugins/browserRouteModulesPlugin_v2";
 import { cssFilePlugin } from "../plugins/cssImports";
-import { deprecatedRemixPackagePlugin } from "../plugins/deprecatedRemixPackagePlugin";
-import { emptyModulesPlugin } from "../plugins/emptyModulesPlugin";
+import { deprecatedRemixPackagePlugin } from "../plugins/deprecatedRemixPackage";
+import { emptyModulesPlugin } from "../plugins/emptyModules";
 import { mdxPlugin } from "../plugins/mdx";
 import { urlImportsPlugin } from "../plugins/urlImportsPlugin";
 // import { cssBundleUpdatePlugin } from "../plugins/cssBundleUpdatePlugin";
