+++ packages/remix-dev/compiler/serverjs/compiler.ts
@@ -12,8 +12,8 @@ import { cssModulesPlugin } from "../plugins/cssModuleImports";
 import { cssSideEffectImportsPlugin } from "../plugins/cssSideEffectImports";
 import { vanillaExtractPlugin } from "../plugins/vanillaExtract";
 import { cssFilePlugin } from "../plugins/cssImports";
-import { deprecatedRemixPackagePlugin } from "../plugins/deprecatedRemixPackagePlugin";
-import { emptyModulesPlugin } from "../plugins/emptyModulesPlugin";
+import { deprecatedRemixPackagePlugin } from "../plugins/deprecatedRemixPackage";
+import { emptyModulesPlugin } from "../plugins/emptyModules";
 import { mdxPlugin } from "../plugins/mdx";
 import { serverAssetsManifestPlugin } from "./plugins/manifest";
 import { serverBareModulesPlugin } from "./plugins/bareImports";
